+++ js/dropdown/shared/Dropdown.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import DropdownOptions from "./DropdownOptions.svelte";
-	import {handle_filter, handle_change} from "./utils"
+	import {handle_filter, handle_change, dispatch_blur, dispatch_select, handle_key_down, handle_focus} from "./utils"
 	import { afterUpdate } from "svelte";
 	import { BlockTitle } from "@gradio/atoms";
 	import { Remove, DropdownArrow } from "@gradio/icons";
@@ -8,10 +8,9 @@
 
 	export let label: string;
 	export let info: string | undefined = undefined;
-	export let value: string | string[] | undefined;
+	export let value: string | undefined;
 	let old_value = Array.isArray(value) ? value.slice() : value;
 	export let value_is_output = false;
-	export let max_choices: number;
 	export let choices: [string, string][];
 	export let disabled = false;
 	export let show_label: boolean;
@@ -48,119 +47,26 @@
 		}
 	}
 
-	/* Handlers for both single-select and multi-select dropdowns */
+	function handle_option_selected(e: any, choices: [string, string][]): void {
+		const option_index = e.detail.target.dataset.index;
+		const [option_name, option_value] = choices[option_index];
+
+		input_text = option_name;
+		value = option_value;
+		show_options = false;
+		
+		dispatch_select(option_index, option_value);
+		filter_input.blur();
+	}
 
 	function handle_blur(): void {
 		if (blurring) return;
 		blurring = true;
-		if (multiselect) {
-			input_text = "";
-		} else if (!allow_custom_value) {
+		if (!allow_custom_value) {
 			input_text = choices_names[choices_values.indexOf(value as string)];
 		}
 		show_options = false;
-		dispatch("blur");
-		setTimeout(() => {
-			blurring = false;
-		}, 100);
-	}
-
-
-	/* Handlers specifically for multi-select dropdown */
-
-	function add(option_index: number): void {
-		value = value as string[];
-		if (!max_choices || value.length < max_choices) {
-			value.push(choices_values[option_index]);
-			dispatch("select", {
-				index: option_index,
-				value: choices_values[option_index],
-				selected: true
-			});
-		}
-		value = value;
-	}
-
-	function remove(option_index: number): void {
-		const option_value = choices_values[option_index];
-		if (!disabled) {
-			value = value as string[];
-			value = value.filter((v: string) => v !== option_value);
-		}
-		dispatch("select", {
-			index: option_index,
-			value: choices[option][1],
-			selected: false
-		});
-	}
-
-	function remove_all(e: any): void {
-		value = [];
-		input_text = "";
-		e.preventDefault();
-	}
-
-	function handle_focus(e: FocusEvent): void {
-		if (blurring) {
-			// Remove focus triggered by blurring to the label.
-			let target = e.target as HTMLInputElement;
-			return target.blur();
-		}
-		dispatch("focus");
-		filtered_indices = choices.map((_, i) => i);
-		show_options = true;
-	}
-
-	// eslint-disable-next-line complexity
-	function handle_key_down(e: any): void {
-		if (e.key === "Enter") {
-			if (!multiselect) {
-				if (value !== activeOption) {
-					value = activeOption;
-					dispatch("select", {
-						index: choices.indexOf(value),
-						value: value,
-						selected: true
-					});
-				}
-				input_text = activeOption;
-				show_options = false;
-				filter_input.blur();
-			} else if (multiselect && Array.isArray(value)) {
-				value.includes(activeOption) ? remove(activeOption) : add(activeOption);
-				input_text = "";
-			}
-		} else {
-			show_options = true;
-			if (e.key === "ArrowUp" || e.key === "ArrowDown") {
-				if (activeOption === null) {
-					activeOption = filtered[0];
-				}
-				const increment = e.key === "ArrowUp" ? -1 : 1;
-				const calcIndex = filtered.indexOf(activeOption) + increment;
-				activeOption =
-					calcIndex < 0
-						? filtered[filtered.length - 1]
-						: calcIndex === filtered.length
-						? filtered[0]
-						: filtered[calcIndex];
-				e.preventDefault();
-			} else if (e.key === "Escape") {
-				show_options = false;
-			} else if (e.key === "Backspace") {
-				if (
-					multiselect &&
-					(!input_text || input_text === "") &&
-					Array.isArray(value) &&
-					value.length > 0
-				) {
-					remove(value[value.length - 1]);
-					input_text = "";
-				}
-			} else {
-				show_options = true;
-			}
-		}
+		dispatch_blur();
 	}
 
 	afterUpdate(() => {
@@ -173,25 +79,6 @@
 
 	<div class="wrap">
 		<div class="wrap-inner" class:show_options>
-			{#if multiselect && Array.isArray(value)}
-				{#each value as s}
-					<!-- TODO: fix -->
-					<!-- svelte-ignore a11y-click-events-have-key-events -->
-					<!-- svelte-ignore a11y-no-static-element-interactions-->
-					<div on:click|preventDefault={() => remove(s)} class="token">
-						<span>{s}</span>
-						{#if !disabled}
-							<div
-								class:hidden={disabled}
-								class="token-remove"
-								title={$_("common.remove") + " " + s}
-							>
-								<Remove />
-							</div>
-						{/if}
-					</div>
-				{/each}
-			{/if}
 			<div class="secondary-wrap">
 				<input
 					class="border-none"
@@ -209,17 +96,6 @@
 					on:blur={handle_blur}
 					on:focus={handle_focus}
 				/>
-				<!-- TODO: fix -->
-				<!-- svelte-ignore a11y-click-events-have-key-events -->
-				<!-- svelte-ignore a11y-no-static-element-interactions-->
-				<div
-					class:hide={!multiselect || !value?.length || disabled}
-					class="token-remove remove-all"
-					title={$_("common.clear")}
-					on:click={remove_all}
-				>
-					<Remove />
-				</div>
 				<DropdownArrow />
 			</div>
 		</div>
@@ -240,7 +116,6 @@
 	label:not(.container) .wrap,
 	label:not(.container) .wrap-inner,
 	label:not(.container) .secondary-wrap,
-	label:not(.container) .token,
 	label:not(.container) input {
 		height: 100%;
 	}
@@ -268,42 +143,6 @@
 		gap: var(--checkbox-label-gap);
 		padding: var(--checkbox-label-padding);
 	}
-
-	.token {
-		display: flex;
-		align-items: center;
-		transition: var(--button-transition);
-		cursor: pointer;
-		box-shadow: var(--checkbox-label-shadow);
-		border: var(--checkbox-label-border-width) solid
-			var(--checkbox-label-border-color);
-		border-radius: var(--button-small-radius);
-		background: var(--checkbox-label-background-fill);
-		padding: var(--checkbox-label-padding);
-		color: var(--checkbox-label-text-color);
-		font-weight: var(--checkbox-label-text-weight);
-		font-size: var(--checkbox-label-text-size);
-		line-height: var(--line-md);
-	}
-
-	.token > * + * {
-		margin-left: var(--size-2);
-	}
-
-	.token-remove {
-		fill: var(--body-text-color);
-		display: flex;
-		justify-content: center;
-		align-items: center;
-		cursor: pointer;
-		border: var(--checkbox-border-width) solid var(--border-color-primary);
-		border-radius: var(--radius-full);
-		background: var(--background-fill-primary);
-		padding: var(--size-0-5);
-		width: 18px;
-		height: 18px;
-	}
-
 	.secondary-wrap {
 		display: flex;
 		flex: 1 1 0%;
@@ -327,16 +166,6 @@
 		-webkit-opacity: 1;
 		opacity: 1;
 		cursor: not-allowed;
-	}
-
-	.remove-all {
-		margin-left: var(--size-1);
-		width: 20px;
-		height: 20px;
-	}
-
-	.hide {
-		display: none;
 	}
 
 	.subdued {
